refine cache service:
mainly to RelationshipService and PostService

+++ model/src/main/java/sjph/life/model/dao/RelationshipDao.java
@@ -33,7 +33,7 @@ public interface RelationshipDao {
      * @param userId
      * @return the number of followers per userId
      */
-    Long getNumberOfFollower(Long userId);
+    int getNumberOfFollower(Long userId);
 
     /**
      * @param userId
